+++ src/Column.py
@@ -1,5 +1,8 @@
+from ColumnType import ColumnType
+
 class Column:
     def __init__(self, id, name, type, dependencies):
+        assert isinstance(type, ColumnType)
         self.id = id
         self.name = name
         self.type = type
